loggator-api/pyproject.toml
[project]
name = "loggator-api"
version = "0.1.0"
description = "AI-powered log analysis tool — OpenSearch + pluggable LLM"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0",
    "boto3>=1.42.91",
    "cryptography>=46.0.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "opensearch-py[async]>=3.1.0",
    "pgvector>=0.4.2",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "pydantic-settings>=2.13.1",
    "requests-aws4auth>=1.3.1",
    "sentry-sdk[fastapi]>=2.0.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
    "typer>=0.24.1",
    "uvicorn[standard]>=0.44.0",
    "langchain-core>=0.3",
    "langchain-ollama>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "python-jose[cryptography]>=3.5.0",
]

[project.scripts]
loggator = "cli.main:app"
